requirements.txt
python-telegram-bot==22.4
pymongo==4.15.1
python-dotenv==1.0.1